worker/pyproject.toml
[project]
name = "fls_worker"
version = "0.1.0"
description = "Dead simple grading worker for 6.S913"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
[project.scripts]
fls-worker = "fls_worker.main:main"
